Horizon kilo-1 milestone (2015.1.0b1)

Conflicts:
	horizon/locale/ca/LC_MESSAGES/django.po
	horizon/locale/ca/LC_MESSAGES/djangojs.po
	horizon/locale/cs/LC_MESSAGES/django.po
	horizon/locale/cs/LC_MESSAGES/djangojs.po
	horizon/locale/es_MX/LC_MESSAGES/django.po
	horizon/locale/fil/LC_MESSAGES/django.po
	horizon/locale/he/LC_MESSAGES/djangojs.po
	horizon/locale/hu/LC_MESSAGES/django.po
	horizon/locale/id/LC_MESSAGES/django.po
	horizon/locale/id/LC_MESSAGES/djangojs.po
	horizon/locale/it/LC_MESSAGES/django.po
	horizon/locale/nl_NL/LC_MESSAGES/django.po
	horizon/locale/nl_NL/LC_MESSAGES/djangojs.po
	horizon/locale/pt/LC_MESSAGES/django.po
	horizon/locale/pt/LC_MESSAGES/djangojs.po
	horizon/locale/sl_SI/LC_MESSAGES/django.po
	horizon/locale/sr/LC_MESSAGES/django.po
	horizon/locale/sr/LC_MESSAGES/djangojs.po
	horizon/locale/te_IN/LC_MESSAGES/djangojs.po
	horizon/locale/tr_TR/LC_MESSAGES/django.po
	horizon/locale/tr_TR/LC_MESSAGES/djangojs.po
	openstack_dashboard/locale/cs/LC_MESSAGES/django.po
	openstack_dashboard/locale/nl_NL/LC_MESSAGES/django.po
	openstack_dashboard/locale/sr/LC_MESSAGES/django.po

+++ openstack_dashboard/api/swift.py
@@ -18,6 +18,7 @@
 
 import logging
 
+from oslo.utils import timeutils
 import six.moves.urllib.parse as urlparse
 import swiftclient
 
@@ -28,7 +29,6 @@ from horizon import exceptions
 from horizon.utils.memoized import memoized  # noqa
 
 from openstack_dashboard.api import base
-from openstack_dashboard.openstack.common import timeutils
 
 
 LOG = logging.getLogger(__name__)
@@ -110,8 +110,6 @@ def swift_api(request):
     endpoint = base.url_for(request, 'object-store')
     cacert = getattr(settings, 'OPENSTACK_SSL_CACERT', None)
     insecure = getattr(settings, 'OPENSTACK_SSL_NO_VERIFY', False)
-    LOG.debug('Swift connection created using token "%s" and url "%s"'
-              % (request.user.token.id, endpoint))
     return swiftclient.client.Connection(None,
                                          request.user.username,
                                          None,
@@ -165,7 +163,8 @@ def swift_get_container(request, container_name, with_data=True):
             swift_endpoint = base.url_for(request,
                                           'object-store',
                                           endpoint_type='publicURL')
-            public_url = swift_endpoint + '/' + urlparse.quote(container_name)
+            parameters = urlparse.quote(container_name.encode('utf8'))
+            public_url = swift_endpoint + '/' + parameters
         ts_float = float(headers.get('x-timestamp'))
         timestamp = timeutils.iso8601_from_timestamp(ts_float)
     except Exception:
@@ -201,8 +200,8 @@ def swift_delete_container(request, name):
     # be done in swiftclient instead of Horizon.
     objects, more = swift_get_objects(request, name)
     if objects:
-        error_msg = unicode(_("The container cannot be deleted "
-                              "since it's not empty."))
+        error_msg = _("The container cannot be deleted "
+                      "since it is not empty.")
         exc = exceptions.Conflict(error_msg)
         exc._safe_message = error_msg
         raise exc
@@ -219,7 +218,7 @@ def swift_get_objects(request, container_name, prefix=None, marker=None,
                   delimiter=FOLDER_DELIMITER,
                   full_listing=True)
     headers, objects = swift_api(request).get_container(container_name,
-                                                          **kwargs)
+                                                        **kwargs)
     object_objs = _objectify(objects, container_name)
 
     if(len(object_objs) > limit):
@@ -308,6 +307,21 @@ def swift_create_pseudo_folder(request, container_name, pseudo_folder_name):
 
 
 def swift_delete_object(request, container_name, object_name):
+    objects, more = swift_get_objects(request, container_name,
+                                      prefix=object_name)
+    # In case the given object is pseudo folder,
+    # it can be deleted only if it is empty.
+    # swift_get_objects will return at least
+    # one object (i.e container_name) even if the
+    # given pseudo folder is empty. So if swift_get_objects
+    # returns more than one object then only it will be
+    # considered as non empty folder.
+    if len(objects) > 1:
+        error_msg = _("The pseudo folder cannot be deleted "
+                      "since it is not empty.")
+        exc = exceptions.Conflict(error_msg)
+        exc._safe_message = error_msg
+        raise exc
     swift_api(request).delete_object(container_name, object_name)
     return True
